new resolve function

+++ app/src/main/java/com/example/agenda/domain/entities/Bank.kt
@@ -1,5 +1,6 @@
 package com.example.agenda.domain.entities
 
+import MONEY
 import com.example.agenda.app.common.Entity
 import io.objectbox.annotation.Entity as E
 import io.objectbox.annotation.Id
@@ -7,11 +8,11 @@ import io.objectbox.annotation.Id
 @E
 data class Bank(
     val name: String,
-    var balance: Float,
+    var balance: MONEY,
     override var id: String?,
     @Id
     var _id: Long = 0,
     override var created_at: Long?,
     override var updated_at: Long?,
-    val credit: Float?,
+    val credit: MONEY?,
 ) : Entity {init {createMetadata()}}
