+++ app/src/main/java/com/example/agenda/domain/entities/Transaction.kt
@@ -1,5 +1,6 @@
 package com.example.agenda.domain.entities
 
+import MONEY
 import com.example.agenda.app.common.Entity
 import com.example.agenda.app.common.RECURRENCE
 import com.example.agenda.app.common.RecurrenceEntity
@@ -17,7 +18,7 @@ data class Transaction(
     val day: Int,
     val month: Int,
     val year: Int,
-    val amount: Float,
+    val amount: MONEY,
     val description: String,
     override var created_at: Long?,
     override var updated_at: Long?,
